% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFUN.R
\name{msiq}
\alias{msiq}
\title{use MSIQ to quantify transcript expression from multiple samples}
\usage{
msiq(D, gtf_path, bam_path, A = 5, B = 2, ncores)
}
\arguments{
\item{D}{An integer specifying the number of RNA-seq sample supplied}

\item{gtf_path}{A character specifying the full path of the .gtf file}

\item{bam_path}{A character vector. Each element gives the full path of one .bam file.}

\item{A}{An integer specifying the first parameter used the Beta prior. Default is 5.}

\item{B}{An integer specifying the second parameter used the Beta prior. Default is 2.}

\item{ncores}{An integer denoting the number of cores used for parallel computation.}
}
\value{
Save the estimation results to a text file transcript_summary.txt.
}
\description{
use MSIQ to quantify transcript expression from multiple samples
}
